backend/requirements.txt
fastapi
uvicorn[standard]
joblib
scikit-learn
pandas
requests
python-dotenv
openai
xgboost
